# app/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "鸵小妥API",
        "contact": {
            "email": "wq0426@163.com"
        },
        "version": "1.0.2"
    },
    "host": "127.0.0.1:8289",
    "paths": {
...
        "/cart/add": {
            "post": {
                "description": "添加商品到购物车，如果已存在则数量加1",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "购物车"
                ],
                "summary": "添加商品到购物车",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "添加购物车请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AddToCartRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            }
        },
        "/cart/delete": {
            "post": {
                "description": "删除购物车中的商品，支持批量删除",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "购物车"
                ],
                "summary": "删除购物车商品",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "删除购物车请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.DeleteCartItemsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            }
        },
        "/cart/list": {
            "get": {
                "description": "获取用户购物车中未处理的商品列表，按店铺分组",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "购物车"
                ],
                "summary": "获取用户购物车列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.CartStoreDTO"
                            }
                        }
                    }
                }
            }
        },
        "/coupon/claim": {
            "post": {
                "description": "根据优惠券ID领取优惠券，如果已领取会提示",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "优惠券"
                ],
                "summary": "领取优惠券",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "领取优惠券请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ClaimCouponRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            }
        },
        "/coupon/detail": {
            "get": {
                "description": "根据优惠券ID获取优惠券详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "优惠券"
                ],
                "summary": "获取优惠券详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "优惠券ID",
                        "name": "coupon_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CouponDetailResponse"
                        }
                    }
                }
            }
        },
        "/coupon/list": {
            "get": {
                "description": "获取用户所有优惠券列表，可根据状态过滤",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "优惠券"
                ],
                "summary": "获取用户所有优惠券",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.ProductCouponDTO"
                            }
                        }
                    }
                }
            }
        },
        "/coupon/product": {
            "get": {
                "description": "获取用户针对特定商品的优惠券列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "优惠券"
                ],
                "summary": "获取用户特定商品的优惠券",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "商品ID",
                        "name": "product_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.UserCoupon"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.AddToCartRequest": {
            "type": "object",
            "required": [
                "product_id",
                "quantity"
            ],
            "properties": {
                "coupon_id": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
...
        "model.CartProductDTO": {
            "type": "object",
            "properties": {
                "cart_id": {
                    "type": "integer"
                },
                "coupon_id": {
                    "type": "integer"
                },
                "coupon_price": {
                    "type": "number"
                },
                "courier_fee_min": {
                    "type": "number"
                },
                "current_price": {
                    "type": "number"
                },
                "member_discount": {
                    "type": "number"
                },
                "product_id": {
                    "type": "integer"
                },
                "product_name": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "model.CartStoreDTO": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.CartProductDTO"
                    }
                },
                "store_id": {
                    "type": "integer"
                },
                "store_name": {
                    "type": "string"
                },
                "store_url": {
                    "type": "string"
                }
            }
        },
        "model.ClaimCouponRequest": {
            "type": "object",
            "required": [
                "coupon_id"
            ],
            "properties": {
                "coupon_id": {
                    "description": "优惠券ID",
                    "type": "integer"
                }
            }
        },
        "model.CouponDetailResponse": {
            "type": "object",
            "properties": {
                "available_min_price": {
                    "description": "可用最低金额",
                    "type": "number"
                },
                "coupon_name": {
                    "description": "优惠券名称",
                    "type": "string"
                },
                "coupon_price": {
                    "description": "优惠券金额",
                    "type": "number"
                },
                "created_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "deadline": {
                    "description": "截止时间",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "product_id": {
                    "description": "商品ID",
                    "type": "integer"
                },
                "status": {
                    "description": "状态",
                    "type": "integer"
                },
                "type": {
                    "description": "类型（1:优惠券；2:兑换券）",
                    "type": "integer"
                }
            }
        },
...
        "model.DeleteCartItemsRequest": {
            "type": "object",
            "required": [
                "ids"
            ],
            "properties": {
                "ids": {
                    "description": "Comma-separated list of cart IDs",
                    "type": "string"
                }
            }
        },
...
        "model.ProductCouponDTO": {
            "type": "object",
            "properties": {
                "available_min_price": {
                    "type": "number"
                },
                "coupon_id": {
                    "type": "integer"
                },
                "coupon_name": {
                    "type": "string"
                },
                "coupon_price": {
                    "type": "number"
                },
                "deadline": {
                    "type": "string"
                },
                "is_received": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "integer"
                }
            }
        },
...
        "model.UserCoupon": {
            "type": "object",
            "properties": {
                "available_min_price": {
                    "description": "可用最低价格",
                    "type": "number"
                },
                "coupon_id": {
                    "description": "优惠券ID (注意:列名在DB中有拼写错误)",
                    "type": "integer"
                },
                "coupon_name": {
                    "description": "优惠券名称",
                    "type": "string"
                },
                "coupon_price": {
                    "description": "优惠券价格",
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "deadline": {
                    "description": "截止时间",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "product_id": {
                    "description": "商品ID",
                    "type": "integer"
                },
                "status": {
                    "description": "状态",
                    "type": "integer"
                },
                "type": {
                    "description": "类型",
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "description": "用户ID",
                    "type": "string"
                }
            }
        },
...
        "v1.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}